+++ spec/Scribunto/EchoTest.lua
@@ -1,0 +1,5 @@
+local p = {}
+function p.echo(frame)
+    return frame.args[1]
+end
+return p