+++ spec/Scribunto/Test2.lua
@@ -1,0 +1,5 @@
+return {
+	countBeans = function ()
+		return 3
+	end
+}